# fix: trie lookup performance