--- a/packages/cspell-lib/src/lib/Models/Suggestion.ts
+++ b/packages/cspell-lib/src/lib/Models/Suggestion.ts
@@ -11,4 +11,9 @@ export interface ExtendedSuggestion {
      * The suggested word adjusted to match the original case.
      */
     wordAdjustedToMatchCase?: string;
+    /**
+     * The cost of using this word.
+     * The lower the cost, the better the suggestion.
+     */
+    cost?: number;
 }