--- a/packages/cspell-trie-lib/api/api.d.ts
+++ b/packages/cspell-trie-lib/api/api.d.ts
@@ -80,6 +80,11 @@ interface TrieInfo {
     forbiddenWordPrefix: string;
     isCaseAware: boolean;
 }
+interface TrieCharacteristics {
+    hasForbiddenWords: boolean;
+    hasCompoundWords: boolean;
+    hasNonStrictWords: boolean;
+}
 type PartialTrieInfo = PartialWithUndefined<TrieInfo> | undefined;
 
 interface FindResult$1 {
@@ -135,7 +140,7 @@ interface ITrieNode {
     findExact?: ((word: string) => boolean) | undefined;
 }
 interface ITrieNodeRoot extends ITrieNode {
-    info: Readonly<TrieInfo>;
+    readonly info: Readonly<TrieInfo>;
     /**
      * converts an `id` into a node.
      * @param id an of a ITrieNode in this Trie
@@ -150,9 +155,12 @@ interface ITrieNodeRoot extends ITrieNode {
      */
     find?: ((word: string, strict: boolean) => FindResult$1 | undefined) | undefined;
     isForbidden?: ((word: string) => boolean) | undefined;
-    forbidPrefix: string;
-    compoundFix: string;
-    caseInsensitivePrefix: string;
+    readonly forbidPrefix: string;
+    readonly compoundFix: string;
+    readonly caseInsensitivePrefix: string;
+    readonly hasForbiddenWords: boolean;
+    readonly hasCompoundWords: boolean;
+    readonly hasNonStrictWords: boolean;
 }
 
 declare const FLAG_WORD = 1;
@@ -374,8 +382,8 @@ declare function suggestionCollector(wordToMatch: string, options: SuggestionCol
  */
 declare function impersonateCollector(collector: SuggestionCollector, word: string): SuggestionCollector;
 
-interface TrieData {
-    info: Readonly<TrieInfo>;
+interface TrieData extends Readonly<TrieCharacteristics> {
+    readonly info: Readonly<TrieInfo>;
     /** Method used to split words into individual characters. */
     wordToCharacters(word: string): readonly string[];
     /** get an iterable for all the words in the dictionary. */
@@ -384,8 +392,10 @@ interface TrieData {
     getNode(prefix: string): ITrieNode | undefined;
     has(word: string): boolean;
     isForbiddenWord(word: string): boolean;
-    hasForbiddenWords(): boolean;
-    size: number;
+    readonly hasForbiddenWords: boolean;
+    readonly hasCompoundWords: boolean;
+    readonly hasNonStrictWords: boolean;
+    readonly size: number;
 }
 
 interface ITrie {
@@ -461,12 +471,16 @@ interface ITrie {
      * On the returned Iterator, calling .next(goDeeper: boolean), allows for controlling the depth.
      */
     iterate(): WalkerIterator;
-    weightMap: WeightMap | undefined;
-    get isCaseAware(): boolean;
+    readonly weightMap: WeightMap | undefined;
+    readonly isCaseAware: boolean;
+    readonly hasForbiddenWords: boolean;
+    readonly hasCompoundWords: boolean;
+    readonly hasNonStrictWords: boolean;
 }
 interface FindWordOptions$1 {
     caseSensitive?: boolean;
     useLegacyWordCompounds?: boolean | number;
+    checkForbidden?: boolean;
 }
 
 declare function buildITrieFromWords(words: Iterable<string>, info?: PartialTrieInfo): ITrie;