--- a/packages/cspell-trie-lib/api/api.d.ts
+++ b/packages/cspell-trie-lib/api/api.d.ts
@@ -117,18 +117,16 @@ interface ITrieNode {
     readonly id: ITrieNodeId;
     /** flag End of Word */
     readonly eow: boolean;
-    /** number of children */
-    readonly size: number;
     /** get keys to children */
-    keys(): readonly string[];
+    keys(): Iterable<string>;
     /** get keys to children */
-    values(): readonly ITrieNode[];
+    values(): Iterable<ITrieNode>;
     /** get the children as key value pairs */
-    entries(): readonly Entry[];
+    entries(): Iterable<Entry>;
     /** get child ITrieNode */
     get(char: string): ITrieNode | undefined;
-    /** get a child by the key index */
-    child(idx: number): ITrieNode;
+    /** get a nested child ITrieNode */
+    getNode?: (chars: string) => ITrieNode | undefined;
     /** has child */
     has(char: string): boolean;
     /** `true` iff this node has children */
@@ -143,7 +141,7 @@ interface ITrieNodeRoot extends ITrieNode {
      * @param id an of a ITrieNode in this Trie
      */
     resolveId(id: ITrieNodeId): ITrieNode;
-    findExact?: ((word: string) => boolean) | undefined;
+    findExact: ((word: string) => boolean) | undefined;
     /**
      * Try to find a word.
      * @param word - the normalized word to look up.