--- a/packages/cspell-trie-lib/src/lib/ITrieNode/FindOptions.ts
+++ b/packages/cspell-trie-lib/src/lib/ITrieNode/FindOptions.ts
@@ -4,7 +4,8 @@ import type { CompoundModes } from './CompoundModes.js';
 export interface FindOptions {
     matchCase: boolean;
     compoundMode: CompoundModes;
-    legacyMinCompoundLength?: number;
+    legacyMinCompoundLength?: number | undefined;
+    checkForbidden?: boolean | undefined;
 }
 
 export type PartialFindOptions = PartialWithUndefined<FindOptions> | undefined;