--- a/packages/cspell-trie-lib/src/lib/ITrieNode/TrieInfo.ts
+++ b/packages/cspell-trie-lib/src/lib/ITrieNode/TrieInfo.ts
@@ -6,4 +6,11 @@ export interface TrieInfo {
     forbiddenWordPrefix: string;
     isCaseAware: boolean;
 }
+
+export interface TrieCharacteristics {
+    hasForbiddenWords: boolean;
+    hasCompoundWords: boolean;
+    hasNonStrictWords: boolean;
+}
+
 export type PartialTrieInfo = PartialWithUndefined<TrieInfo> | undefined;