--- a/packages/cspell-trie-lib/src/lib/constants.ts
+++ b/packages/cspell-trie-lib/src/lib/constants.ts
@@ -12,4 +12,7 @@ export const defaultTrieInfo: TrieInfo = Object.freeze({
     forbiddenWordPrefix: FORBID_PREFIX,
     stripCaseAndAccentsPrefix: CASE_INSENSITIVE_PREFIX,
     isCaseAware: true,
+    hasForbiddenWords: false,
+    hasCompoundWords: false,
+    hasNonStrictWords: false,
 });