… .github/workflows/integration-test.yml
… packages/cspell-lib/src/lib/suggestions.test.ts
… packages/cspell-lib/src/lib/textValidation/__snapshots__/docValidator.test.ts.snap
… packages/cspell-lib/src/lib/textValidation/docValidator.test.ts
… packages/cspell-tools/package.json
… packages/cspell-trie-lib/src/lib/ITrieNode/find.test.ts
… packages/cspell-trie-lib/src/lib/TrieBlob/FastTrieBlob.ts
… packages/cspell-trie-lib/src/lib/TrieBlob/FastTrieBlobIRoot.test.ts
… packages/cspell-trie-lib/src/lib/TrieBlob/FastTrieBlobInternals.ts
… packages/cspell-trie-lib/src/lib/TrieBlob/TrieBlob.ts
… packages/cspell-trie-lib/src/lib/TrieBlob/TrieBlobIRoot.test.ts
… packages/cspell-trie-lib/src/lib/TrieBlob/TrieBlobIRoot.ts
… packages/cspell-trie-lib/src/lib/TrieData.ts
… packages/cspell-trie-lib/src/lib/TrieNode/TrieNodeTrie.ts
… packages/cspell-trie-lib/src/lib/TrieNode/find.ts
… packages/cspell-trie-lib/src/lib/TrieNode/trie.ts
… packages/cspell-trie-lib/src/lib/buildITrie.ts
… packages/cspell-trie-lib/src/lib/io/importV3.test.ts
… packages/cspell-trie-lib/src/lib/io/importV3FastBlob.test.ts
… packages/cspell-trie-lib/src/lib/suggestions/suggestAStar.test.ts
… packages/cspell-trie-lib/src/lib/suggestions/suggestAStar.ts
… packages/cspell-trie-lib/src/lib/suggestions/suggestTrieData.test.ts
… packages/cspell-trie-lib/src/lib/walker/walker.ts
… packages/cspell-trie-lib/src/perf/charIndex.perf.ts
… packages/cspell-trie-lib/src/perf/misc.perf.ts
… packages/cspell/src/app/__snapshots__/app.test.ts.snap
… pnpm-lock.yaml