This is currently a place holder for per work.

Initial Perf:
```
File: src/perf/has.perf.ts
Running Perf Suite: trie has
✔ trie has words             7.31 ops/sec      4 iterations  546.87ms time
✔ fastTrieBlob has words    14.43 ops/sec      8 iterations  554.57ms time
✔ trieBlob has words        19.27 ops/sec     10 iterations  518.86ms time
✔ iTrieFast has words       12.64 ops/sec      7 iterations  553.83ms time
✔ iTrieBlob has words       16.79 ops/sec      9 iterations  536.03ms time
done.
File: src/perf/has.perf.ts
Running Perf Suite: dictionary has
✔ dictionary has 100k words               24.43 ops/sec     13 iterations  532.14ms time
✔ dictionary has 100k words (2nd time)    24.68 ops/sec     13 iterations  526.69ms time
✔ collection has 100k words               13.44 ops/sec      7 iterations  520.71ms time
✔ iTrie has 100k words                    32.54 ops/sec     17 iterations  522.51ms time
✔ iTrie.hasWord has 100k words            33.18 ops/sec     17 iterations  512.32ms time
✔ iTrie.data has 100k words               37.27 ops/sec     19 iterations  509.79ms time
Running Perf Suite: dictionary has Not
✔ dictionary has not 100k words                6.65 ops/sec      4 iterations  601.33ms time
✔ dictionary has not 100k words (2nd time)     6.36 ops/sec      4 iterations  628.84ms time
✔ collection has not 100k words                2.57 ops/sec      2 iterations  776.87ms time
✔ iTrie has not 100k words                    29.93 ops/sec     15 iterations  501.20ms time
✔ iTrie.hasWord has not 100k words            27.78 ops/sec     14 iterations  503.89ms time
✔ iTrie.data has not 100k words               33.93 ops/sec     17 iterations  500.99ms time
done.
```

result:
```
File: src/perf/has.perf.ts
Running Perf Suite: trie has
✔ trie has words             6.44 ops/sec      4 iterations  620.65ms time
✔ fastTrieBlob has words    13.74 ops/sec      7 iterations  509.65ms time
✔ trieBlob has words        19.24 ops/sec     10 iterations  519.73ms time
✔ iTrieFast has words       13.90 ops/sec      7 iterations  503.53ms time
✔ iTrieBlob has words       16.88 ops/sec      9 iterations  533.31ms time
done.
File: src/perf/has.perf.ts
Running Perf Suite: dictionary has
✔ dictionary has 100k words            25.89 ops/sec     13 iterations  502.22ms time
✔ collection has 100k words            13.81 ops/sec      7 iterations  506.81ms time
✔ collection reverse has 100k words    13.26 ops/sec      7 iterations  527.84ms time
✔ iTrie has 100k words                 36.84 ops/sec     19 iterations  515.72ms time
✔ iTrie.hasWord has 100k words         33.99 ops/sec     18 iterations  529.51ms time
✔ iTrie.data has 100k words            38.55 ops/sec     20 iterations  518.75ms time
Running Perf Suite: dictionary has Not
✔ dictionary has not 100k words        6.37 ops/sec      4 iterations  627.52ms time
✔ collection has not 100k words        2.71 ops/sec      2 iterations  738.63ms time
✔ iTrie has not 100k words            28.77 ops/sec     16 iterations  556.20ms time
✔ iTrie.hasWord has not 100k words    25.99 ops/sec     13 iterations  500.14ms time
✔ iTrie.data has not 100k words       34.15 ops/sec     18 iterations  527.10ms time
done.
```